Support clients using CMake (#34)
* Support clients using CMake

Supporting clients using CMake requires that a project provides a CMake
package that can be found via find_package().

The package name was first suggested to be MicrosoftIFC in #29, but to
follow the pattern of GSL I changed it to Microsoft.IFC, which is still
very similar.

The package exports 3 targets:

- Microsoft.IFC::Core
- Microsoft.IFC::DOM
- Microsoft.IFC::SDK

Microsoft.IFC::SDK is just an interface target, which links to the other
targets as a convenient all-in-one target. This enables usage like so:

    find_package(Microsoft.IFC REQUIRED)
    target_link_libraries(project_target PRIVATE Microsoft.IFC::SDK)

The ifc-printer executable was made an opt-in target.

The project now also has an opt-in developer mode option for its
developers via the DEVELOPER_MODE cache variable, which will enable the
printer target and make testing available.

For people who wish not to use vcpkg to fetch GSL, an opt-in find module
is provided to download and setup GSL for the project by providing the
cmake/find directory via the CMAKE_MODULE_PATH command line option. This
can be done depending on your shell like so:

    # cmd
    -D "CMAKE_MODULE_PATH:PATH=%cd%\cmake\find"
    # pwsh & POSIX shell
    -D "CMAKE_MODULE_PATH:PATH=$PWD/cmake/find"

Fixes: #29

* Update BUILDING.md

Word smith MSVC compiler usage

* Update CMakePresets.json

Remove `-Woverloaded-virtual` from GCC and Clang flags.  Guarding against overriding accidentally being overloading is better enforced though C++ Core Guidelines Rule [C.128](https://isocpp.github.io/CppCoreGuidelines/CppCoreGuidelines#Rh-override)

* Update HACKING.md

Word smith the notes in **Presets** section.

---------

Co-authored-by: friendlyanon <friendlyanon@users.noreply.github.com>
Co-authored-by: Gabriel Dos Reis <GabrielDosReis@users.noreply.github.com>